requirements.txt
requests
google-cloud-monitoring
google-cloud-logging